+++ crates/cedra/src/test/tests.rs
@@ -113,11 +113,11 @@ async fn ensure_every_command_args_work() {
 /// Ensure we can parse URLs for args
 #[tokio::test]
 async fn ensure_can_parse_args_with_urls() {
-    let result = ArgWithType::from_str("string:https://aptoslabs.com").unwrap();
+    let result = ArgWithType::from_str("string:https://cedra.network").unwrap();
     matches!(result._ty, FunctionArgType::String);
     assert_eq!(
         result.arg,
-        bcs::to_bytes(&"https://aptoslabs.com".to_string()).unwrap()
+        bcs::to_bytes(&"https://cedra.network".to_string()).unwrap()
     );
 }